+++ lib/maven/goal/mavenTest.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018 Atomist, Inc.
+ * Copyright © 2019 Atomist, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -56,9 +56,11 @@ function executeMavenTest(goalInvocation: GoalInvocation): Promise<{ code: numbe
         {id, credentials, readOnly: true},
         async (p: LocalProject) => {
         const mavenCommand = await determineMavenCommand(p);
-        const result = await spawnLog(
-            mavenCommand, ["test"],
-            {cwd: p.baseDir, log: new LoggingProgressLog("maven-test", "info")});
+        const result = await spawnAndWatch(
+            {command: mavenCommand, args: ["test"]},
+            {cwd: p.baseDir},
+            new LoggingProgressLog("maven-test", "info"),
+            {});
         const r = await getJUnitTestResults(p);
         const prefix = r.tests === 1 ? `1 test` : `${r.tests} tests`;
         if (result.code === 0) {